Using macros for diplomatic option checks.
A lot more could be done, but I think this is a good start.
I think diplomacy should not work with integer-indexed variables though.
We should probably have variables named after units (or something
similar) and possibly an index variable or something. No idea how to do
that in WML.