rearange main.cfg to avoid reading macros multiple times
which spams currently the log

The modification is read independent of MULTIPLAYER.
In the MULTIPLAYER part, the files are now listed explicitely,
to exclude the ones already read from reading.

The parser reads the file from top to down.
It needs to know the macros which are read in the modification
when it reads the MULTIPLAYER part.
To archieve that, the modification is moved to the top of the file.