The bug fix
The  bug was in anl.offer_agriculture / anl.research_field

It was correctly reading the »farming« value in anl.research_field.
But it was written in a way that the name of the value is not only used for that,
but also when changing the research target.

However, there the correct name is agriculture.
Changing the WML side of it to be the same.

This problem only occured after changing the research field to sth.
else and back, because at game start the correct value was set.