remove workaround for hiding units in the help
In 1.14, if one starts a non-ANL game, and then a normal one,
it will still show the changes from ANL in the help.
The new tag [modify_unit_type] allowed to work around this.
However, du to another change the unirs from ANL will not be
available in the non ANL game anymore, so this is fixed in 1.16.