Team coloring for drake images
Formerly the image files were scaled down.
Part of the magengta image slightly changed the color and was not detected as team color.

Now the images are not scaled down, but have the baldes removed in the image files like before.
The scaling is done by the engine after appyling team coloring.